move events from FEF components to fef-events.js
SRFCMSAL-2788
