+++ source/assets/js/classes/fef-dom-observer.js
@@ -1,12 +1,11 @@
+import { DOM_CHANGED_EVENT } from '../utils/fef-events';
+
 $(document).ready(function () {
     let observer = new FefDomObserver();
     observer.triggerDomChangedEvent();
 });
 
-export const DOM_CHANGED_EVENT = 'fef.dom.changed';
-
 export class FefDomObserver {
-
     /**
      * This event can be triggered to react to DOM changes
      */
