+++ source/assets/js/components/fef-form-field.js
@@ -1,5 +1,5 @@
-import {DOM_CHANGED_EVENT} from '../classes/fef-dom-observer';
-import {KEYCODES} from '../utils/fef-keycodes';
+import { KEYCODES } from '../utils/fef-keycodes';
+import { DOM_CHANGED_EVENT } from '../utils/fef-events';
 
 $(window).on(DOM_CHANGED_EVENT, (e) => {
     $('.radio-button, .checkbox').each((index, element) => {
