+++ source/assets/js/components/fef-modal.js
@@ -1,13 +1,12 @@
-import { DOM_CHANGED_EVENT, FefDomObserver } from '../classes/fef-dom-observer';
 import { FefResponsiveHelper } from '../classes/fef-responsive-helper';
 import { FefBouncePrevention } from './fef-bounce-prevention';
 import { KEYCODES } from '../utils/fef-keycodes';
+import { DOM_CHANGED_EVENT, DOM_HEIGHT_CHANGE_EVENT } from '../utils/fef-events';
 
 const ANIMATION_FADE_IN_OUT = 'fade-in-out';
 const ANIMATION_SCALE_FROM_ORIGIN = 'scale-from-origin';
 const ANIMATION_FLYOUT = 'as-flyout-from-origin';
 const ANIMATION_SLIDE_FROM_BOTTOM = 'slide-from-bottom';
-const DOM_HEIGHT_CHANGE_EVENT = 'fef.element.height.changed'; // will be thrown by some element that changed its height
 
 const ANIMATION_SPEED = (window.matchMedia('(prefers-reduced-motion)').matches) ? 0 : 200;
 const END_OF_MODAL = '.js-end-of-modal';
@@ -16,9 +15,9 @@ const END_OF_MODAL = '.js-end-of-modal';
 let existingModals = {};
 let scrollbarWidth = 0;
 
-$(window).on(DOM_CHANGED_EVENT, (e) => {
+$(window).on(DOM_CHANGED_EVENT, () => {
     scrollbarWidth = window.innerWidth - document.documentElement.clientWidth;
-    $('[data-modal-id]').each((index, element) => {
+    $('[data-modal-id]').each((_, element) => {
 
         $(element).off('click').on('click', (event) => {
             event.preventDefault();
